refactor: split functions.php
