+++ messages.php
@@ -3,6 +3,8 @@
 require_once 'init.php';
 require_once 'helpers.php';
 require_once 'functions.php';
+require_once 'db_requests.php';
+require_once 'validation.php';
 
 check_page_access();
