+++ popular.php
@@ -3,6 +3,7 @@
 require_once 'init.php';
 require_once 'helpers.php';
 require_once 'functions.php';
+require_once 'db_requests.php';
 
 check_page_access();
